requirements.txt
streamlit
transformers
langchain
openai
pinecone-client
pdfplumber
python-dotenv

